refactor(api): key queries by account identity and consolidate type exports
